data/requirements.txt
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
sentence-transformers>=3.0.0
qdrant-client>=1.17.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-httpx>=0.30.0
